% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COPD_Health_Index_function.R
\name{COPD_HEALTH_INDEX_fct}
\alias{COPD_HEALTH_INDEX_fct}
\title{Variables having an impact on a disease presence/absence}
\usage{
COPD_HEALTH_INDEX_fct(y, X, signiThresh, slopeThresh)
}
\arguments{
\item{y}{The presence/absence value for the study disease.}

\item{signiThresh}{The pvalue limit of the glm output that we conserve the variables.}

\item{slopeThresh}{The slope value limit of the glm output that we conserve the variables.}

\item{x}{The variables values we want to analyse.}
}
\value{
List of the variables having a pvalue lower thant signiThresh value and a slope value higher than the slopeThresh.
}
\description{
Create a list of the variable having an effect on the presence/absence of a disease. Using a logistic regression for every unique variable in the dataset.
}
\examples{
list1 <- COPD_HEALTH_INDEX_fct(y2,x2,signiThresh = 0.05, slopeThresh = 0.1);
list2 <- COPD_HEALTH_INDEX_fct(y2,x2,signiThresh = 0.05, slopeThresh = 0.1);
}
