% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COPD_Health_Index_function.R
\name{MPOC_OUI_Procruste}
\alias{MPOC_OUI_Procruste}
\title{Procruste analysis for patient with a disease between 2 time-step.}
\usage{
MPOC_OUI_Procruste(x)
}
\arguments{
\item{x}{Dataset with the variable we want to analyse with the individual number and the the disease presence/absence in the first two columns.}
}
\value{
A plot of the distance between the indivuduals and the value on the distance for each of them.
}
\description{
Does a procruste analysis on the patient between two time-step to get a distance of how much their explanatory variables changes between the them.
}
\examples{
Proc1 <- MPOC_OUI_Procrustre(x1);
Proc2 <- MPOC_OUI_Procrustre (x2);
}
